requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pydantic==2.10.5
pydantic-settings==2.7.1
numpy==2.2.1
scipy==1.14.1
scikit-learn==1.6.0
pandas==2.2.3
networkx==3.4.2
PyMuPDF==1.25.1
python-docx==1.1.2
python-multipart==0.0.20
matplotlib>=3.8.0
openai>=1.60.0
azure-storage-blob>=12.24.0
azure-data-tables>=12.6.0
python-dotenv>=1.0.1
requests>=2.31.0
